refactor: 공통으로 적용할 부분 반영(React.FC 지양, const로 선언, 파일명 kebab-case)
