+++ apps/commerce-client-web/src/app/(auth)/join/component/completion-step.tsx
@@ -4,12 +4,12 @@ interface CompletionStepProps {
   userName: string;
 }
 
-const CompletionStep: React.FC<CompletionStepProps> = ({ userName }) => {
+const CompletionStep = ({ userName }: CompletionStepProps) => {
   return (
     <div>
       <h2 className="mb-4 text-2xl font-bold">가입완료</h2>
       <div className="flex h-80 flex-col items-center justify-center">
-        <HandThumbUpIcon className="size-28 text-primary" />
+        <HandThumbUpIcon className="text-primary size-28" />
         <p className="mt-6 text-lg">
           환영합니다, <span className="font-bold">{userName}</span>님!
         </p>
